splitting works

+++ include/ui/button_constants.h
@@ -26,6 +26,7 @@
     + (i % 5) * (CHIP_BUTTON_WIDTH + CHIP_BUTTON_SEPARATION_X)
 #define CHIP_BUTTON_Y(i) \
     CHIP_BUTTON_ORIGIN_Y \
-    + (i / 5) * (CHIP_BUTTON_HEIGHT + CHIP_BUTTON_SEPARATION_Y) \
+    + (int)(i / 5) * (CHIP_BUTTON_HEIGHT + CHIP_BUTTON_SEPARATION_Y) \
     - 30 * ABS(2 - (i % 5)) \
     + 60
+// (int) is unneccesary but it makes clang shut up
